Fix duplicate functions
